refactor(react) : App.jsx 분리
