+++ .gitignore
@@ -31,3 +31,5 @@ libs
 libs/*
 
 !.env.example
+
+GEMINI.md
